Adapt provider local and local end-to-end tests to single-zone istio ingress gateways.
Using multiple istio ingress gateways as entry points in the local setup is challenging as there
are multiple assumptions in the code about how the setup works. The most important one is that
the load balancer is exposed via 127.0.0.1 and the domain name of shoot api servers resolve to
this address so that the standard port (443) can be used for https. To allow the local setup to
continue to work without a real dns server multiple steps had to be taken to adapt the setup.
First of all, the additional single-zone istio-ingress gateways are exposed via 127.0.0.10|11|12
for zone 0, 1 and 2 respectively. The corresponding load balancer services and the host port
mappings are adapted accordingly.
As the domain name of the shoot api server can no longer statically be mapped via /etc/hosts to
127.0.0.1 the coredns server in namespace 'gardener-extension-provider-local-coredns' is exposed
to the host as well. The rewriting logic in the dns configuration is changed to a proper rewrite
per shoot cluster, i.e. during dnsrecord reconciliation provider local adds/removes entries in
the custom dns configuration (analogous to the gardener custom dns configuration).
To allow the host to get a response for a shoot api server domain name, we use a new feature of
core dns 1.10 called view plugin. It allows to provide different results based on expressions
on top of the request. For simplicity, the external requests are assumed to be via tcp while the
cluster internal dns requests are assumed to be via udp. This allows the host to get
127.0.0.1|10|11|12 while the cluster internal requests are resolved to 10.*.*.*.
Last but not least, the test timeouts needed to be adapted as the change in number of replicas of
the istion ingress gateways in conjunction with a slow rollout caused tests to frequently exceed
the previous timeouts.