Adapt multi availability zone handling of istio to local mac requirements.
Using docker on mac can have some implications with regards to port mappings
and local address handling. This change addresses two issues, which prevented
start-up of local kind environment after #6997.
First of all, the DNS server port exposed to be used in the multi-availability
zone setup is moved above the privileged port boundary (1024). UDP support is
removed from the port mapping as it is not required and apparently docker on
mac has issues mapping two protocols on the same port.
In addition to that, the local IP addresses for the zone specific istio ingress
gateways are now assigned as part of the kind start-up script unless they are
already assigned.