[operator] Manage Virtual Garden `kube-apiserver` Exposure (#7238)
* Reconcile virtual cluster `kubeapiserverexposure` component

Co-Authored-By: Oliver Götz <47362717+oliver-goetz@users.noreply.github.com>

This includes the load balancer service for the kube-apiserver exposure.

* Expose virtual garden locally

In the local setup case, the `provider-local`'s `Service` controller is added to the `gardener-operator`'s controller-runtime manager.
This way, the code for reconciling the `Service` resources can be shared/reused.

* Extend `Garden` API with LB annotation settings

Similar to the `.spec.settings.loadBalancerServices.annotations` in the `Seed` API