Drop assumptions about "old service account secrets"
Earlier, the code implicitly assumed that all secrets except the first on in the `ServiceAccount.secrets[]` list are old.
However, this is not true when the user interferes while the rotation is running.
This commit improves the code to REALLY check whether the secret is old by comparing its creation timestamp with the timestamp when the last rotation initiation has finished. All secrets created after this time have definitely been signed with the new key, all others not or not necessarily.