Block access to the Prometheus Management and Targets API
The Prometheus Management API allows to reload the Prometheus configuration or
to quit the Prometheus process.

The shoot control plane Prometheus is also connected to the seed's API server,
so that frequent restarts could put an extra load on the seed API server due to
the service discovery mechanism of Prometheus.

We do not want to block the /-/ready API because it is used by the Prometheus
UI. So we block the sensitive management APIs explicitly. (-> reload and quit)

See https://prometheus.io/docs/prometheus/latest/management_api/

The Prometheus Targets API returns the list of targets that have been discovered
by the Prometheus service discovery mechanism. This could leak information about
the pods that are running in the seed cluster and it is not relevant for shoot
owners. Access to the targets API is blocked as well.

We use the status code 403 Forbidden instead of 401 Unauthorized. In some
browsers, the 401 Unauthorized response leads to showing the password prompt
popup window again in the Prometheus UI, which impacts the usability.

Note that this changes the status code for anonymous requests from 401 (before
this PR) to 403. We couldn't configure nginx to only return 403 after a
successful authentication.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>