[GEP-21] Extend `Shoot`, `Seed`, and `Network` APIs (#7288)
* Add `IPv6SingleStack` feature gate

* Enable `IPv6SingleStack` feature gate in local setup

* Add `ipFamilies` field in `Shoot` API

* Add `ipFamilies` field in `Seed` API

* Add `ipFamilies` field in `Network` API

* Prefactor: unit test the actual overall defaulting func

We should always test the whole thing that is registered as the
defaulting func in the scheme instead of testing only parts of it.

* Default `ipFamilies` to `["IPv4"]`

* Validate basic cases for `ipFamilies` fields

* Validate CIDR values match primary IP family

* Add `IsIPv?SingleStack` helper funcs

This is added to the API package itself instead of to the helper package, because we
can't use the helper package in defaulting (otherwise, we would get a cyclic dependency)

* Default `nodeCIDRMaskSize` according to `ipFamilies`

* Adapt cross-validation of `podCIDR` with `nodeCIDRMaskSize`

Functionally, nothing changed in the validation logic.
However, the current calculation will overflow when using `nodeCIDRMaskSize` values bigger than 32.
In IPv6, `nodeCIDRMaskSize` might be much larger, so we need to use `big.Int` for this calculation.

* Prefactor: rework unit tests for `nodeCIDRMaskSize` validation

* Adapt cross-validation of `nodeCIDRMaskSize` with `maxPods`

Similar to the previous change in the validation code, we need to cater for higher
powers of 2 when `nodeCIDRMaskSize` values increase with IPv6.
The calculation is switched to big.Int, though the logic doesn't change.

* Only validate boundaries of `nodeCIDRMaskSize` for IPv4 single-stack clusters

The valid values for `nodeCIDRMaskSize` on IPv6 clusters is a bit trickier.
It's probably not worth it to duplicate the validation of kube-controller-manager
in our API logic (as it depends on multiple factors, that are hard to cross-validate).

* Address review suggestions (1)

* Drop `pkg/features` from allowed prefixes in .import-restrictions

This package was added to the allowed prefixes during an earlier stage of the PR
where the code checking the feature gate was organized a bit differently.
Now, it's not needed anymore. So drop the allowed prefix everywhere.

* Add unit tests for `DeterminePrimaryIPFamily`

* Add unit tests for `ValidateCIDRIPFamily`

* Make `ValidateNodeCIDRMaskWithMaxPod` more readable

* Replace `HaveLen(0)` with `BeEmpty()`

* Address review suggestions (2)

* Address review suggestions (3)

* Address review suggestions (4)

* Make doc strings for IPFamilies clearer

* Address review suggestions (5)

Co-authored-by: Niclas Schad <niclas.schad@stackit.de>