Add a script to export the state relevant for this PR
We use the development version of the (remote) local setup of gardener as
documented in `docs/development/getting_started_locally.md` to deploy the base
commit of this PR, create a shoot cluster and export the state of the shoot
cluster's control plane that is relevant for this PR.

The `example/provider-local/shoot.yaml` shoot spec is configured with
the alerting feature according to https://gardener.cloud/docs/gardener/monitoring/alerting/
to deploy the alertmanager in the control plane.

Then, we can upgrade to this PR and restart the gardenlet process, let the
changes be applied, and export the relevant state again.

When reviewing this PR, the diff in the exported state as captured in the
consecutive commits might help to go through the major changes and expectations
that are introduced in this PR.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>