"Rename" `NetworkPolicy`: `allow-to-seed-apiserver` -> `allow-to-runtime-apiserver` (#7389)
* Add `gardener.cloud/role=istio-{system,ingress}` label to all `istio-*` namespaces

When the seed has > 1 zone, then there are multiple namespaces running `istio-ingressgateway`s. We need a way to commonly identify them (the names are not static but depending on the concrete zone names).
Also, for consistency, let's also add a role label for the `istio-system` namespace.

* Drop prefix from `SeedAPIServerNetworkPolicyControllerConfiguration`

In the future, this controller will reconcile also other network policies.

* Prefactor `NetworkPolicy` controller

Gardenlet's `NetworkPolicy` controller will become responsible for managing more policies (not only `allow-to-seed-apiserver` like today). Hence, let's prefactor its implementation so that it is ready for this change.

* "Rename" `allow-to-seed-apiserver` policy to `allow-to-runtime-apiserver`

* Adapt components to use `to-runtime-apiserver` instead of `to-seed-apiserver`

* Improve `reconciler.MapToNamespaces` function

Instead of looping over all configs and selectors and then listing namespaces multiple times, we now list all namespaces once and then filter them with a single iteration.

* Address PR review feedback