Adjust the logging-and-monitoring extension documentation
Note that the `dashboard_{users,operators}` fields were marked as deprecated by
mistake some time ago, so the deprecation comment is now removed.

To avoid breaking the extension contract, the field names are not adjusted.
Dashboards that are contributed as either dashboard_operators or dashboard_users
will both be provisioned to the single Grafana instance.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>
Co-authored-by: Kristian-ZH <k.zhelyazkov@sap.com>