Update testmachinery (load) tests for Loki
There are 2 tenants in loki: user and operator, and they were accessible by the
user and operator Grafana instances.

Now the new (user) Grafana uses the "operator" tenant in Loki, and the "user"
tenant in Loki is not used anymore by Grafana.

This "Grafana unification" PR however does not touch the multi-tenancy feature
of Loki. So this commit does not change the expectations of the test, only the
step how the Loki tenant name was extracted from the Grafana ingress annotation
is adjusted.

We expect to find the "operator" tenant in the ingress annotation of Grafana. We
hard code the "user" tenant for the tests because there is no more Grafana
ingress with the user tenant.

Later, the multi tenancy feature can be removed from Loki. This is out of scope
of this PR.

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>
Co-authored-by: Vladimir Vasilev <38783428+vlvasilev@users.noreply.github.com>