Cleanup of old/unusable istio configuration flags/options.
Some of the command line flags and configuration options were already deprecated when the SNI support was
first introduced in gardener. They were in parts removed with the first istio update (1.6). Some other
flags/options were only removed later.
Overall, there should be no change with regards to functionality as the command line flags were simply
ignore. The same was true for the options. The only change was the move of the connectTimeout. However,
this is still the default value. Hence, it should not affect clusters in any way.
(Istio would have complained about some of the options when set to debug log level.)