Add a size limit for the Grafana storage emptyDir volume
The size limit serves as a security precaution.

If a malicious end user could add and save very large dashboards to Grafana or
cause Grafana to write to disk in any other way, it could eventually fill up the
node's disk. This could impact other pods running on the node.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>