Remove the role label from the Grafana helm chart
Previously, the Grafana helm chart was used to render two different Grafana
instances that were distinguished by the role label and the role suffix in the
name of the different artifacts: the users and the operators Grafana.

The role label and name suffix is removed. The Grafana helm chart can render
only a single Grafana instance now.

The non trivial changes will follow in separate commits.

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>