The new user Grafana provisions all the dashboards
Previously, the user Grafana showed the only dashboards from the owners folder,
the operator Grafana showed the dashboards from both the owners and the
operators folders.

Now the new (user) Grafana still provisions the dashboards from the owners
folder, but all the dashboards are stored now in the owners folder.

Hence the new (user) Grafana will show all the dashboards.

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>