Clean up the former Grafana instances (users and operators Grafana)
If a shoot existed before this PR, than two Grafana instances were
deployed to its control plane. These instances have to be deleted
as part of the migration process.

Note that the monitoring component is not yet using managed resources
to deploy its artifacts, so we need to explicitly delete the old
artifacts.

Co-authored-by: Wesley Bermbach <wesley.bermbach@sap.com>
Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>