The grafana-operators deployment, ingress and secret are deleted
Previously, there were 2 Grafana deployments in each control plane:

- grafana-operators that Gardener operators had access to
- grafana-users that shoot owners had access to

Grafana-users contained only a subset of the dashboards that were present in the
grafana-operators instance and shoot owners had restricted access to the Grafana
UI, e.g. could not check the panel definitions, make temporary changes to
dashboards or use the Explore feature of Grafana.

With this PR "Unify the user and operator Grafana deployments" the
grafana-operators instance is deleted and there will be only a single Grafana
deployment in each control plane so that both Gardener operators and shoot
owners will have the same access to observability data.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>