The URL prefix of the grafana ingress does not change
The grafana-users ingress used to have the "gu-" URL prefix which is kept for
the new grafana instance. This way the Grafana URL does not change for shoot
owners due to this PR.

Note that the TLS secret is regenerated during the reconciliation, because the
list of the domains changed: "gu-", "go-" --> "gu-".

Also note that the "operator" Loki tenant is used instead of the "user" Loki
tenant to be able to show all the logs in Grafana.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>