The shoot owner credentials to access Grafana do not change
The grafana-users ingress was configured to use basic authentication. The new
grafana ingress uses the same password for authentication. These credentials are
not rotated implicitly as a side effect of this PR.

Shoot owners can use the same URL with the same credentials to access the new
Grafana instance.

This reduces the complexity of the changes in the Gardener Dashboard repository.
Otherwise, some extra logic would be needed in the Gardener dashboard to provide
shoot owners a valid Grafana link and credentials even during the "transition
period": the time between the new version of the Gardener dashboard is deployed
centrally and the specific shoot cluster's control plane is upgraded in its
maintenance time window some time later.

Note that the secrets did not change when the PR was deployed and hence there is
no diff in these files. To show the relevant files in this git commit, a new
line with a . was added. The hash is used to avoid committing even local secrets
to the git repository.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>