The new grafana instance contains all the dashboards
Previously, the dashboards of the grafana-users instance were only a subset of
the dashboards of the grafana-operators instance.

Shoot owners had access to fewer dashboards when compared to Gardener operators.

The new grafana instance contains all the dashboards of the former
grafana-operators instance and hence shoot owners and Gardener operators can
access the same set of dashboards.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>