The Prometheus UI is accessible to shoot owners
Accessing the Prometheus UI was previously restricted to Gardener operators.
With this PR, shoot owners can also access the Prometheus UI.

The intent of this PR is to provide both for Gardener operators and shoot owners
the same access to the observability data of the control plane.

Technically this is achieved by using the Grafana credentials for the Prometheus
ingress so that shoot owners can access Prometheus with the same password as
Grafana, just like Gardener operators did before. Note that eventually the
password based authentication should be replaced with a generic SSO solution.

Shoot owners can now access both the Prometheus UI and the Prometheus API. They
can federate all or a subset of the data from the control plane Prometheus into
their own monitoring stack, to visualize, alert on the control plane metrics and
correlate them with other observations. This way they can also control the
retention period of these time series. Later, as an alternative to federation
via the password protected ingress, we plan to add a remote write configuration
to Prometheus via the shoot spec to let Prometheus push the requested time
series to an end point that is provided by the shoot owners.

Note that parts of the management API of Prometheus would allow to put
additional load on the seed's API server (by restarting Prometheus and hence
restarting some watches) and could leak information via the service discovery
output (by listing resources in the garden namespace) so these APIs are blocked
on the ingress level.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>