Access to the Alertmanager UI is not changed
If shoot owners configured alerting via the shoot spec:
https://gardener.cloud/docs/gardener/monitoring/alerting/
an alertmanager instance was deployed in the control plane.

Shoot owners could access this alertmanager UI and that aspect is not changed by
this PR: the Grafana (users) password is still configured in the alertmanager
ingress.

To be consistent with the Prometheus ingress configuration, the unneeded
management API access is blocked so that shoot owners can not trigger a reload
of the alertmanager configuration.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>