Allow vpn-client in kube-apiserver to talk to seed apiserver in case of HA VPN.
In the HA VPN, the ip address management of the vpn clients on the seed is done via annotations.
Therefore, vpn-client container of the kube-apiserver needs to be able to communicate with the
seed's apiserver. In case the seed is a gardener shoot this most likely causes vpn-client to
use the external address of the seed's apiserver. This is due to the apiserver-proxy pod mutator
injecting the KUBERNETES_SERVICE_HOST environment variable with the external name.
However, in other seed scenarios vpn-client might also connect to the seed's apiserver via
other means, e.g. through in-cluster communication. For this general case, the existing network
policies were not sufficient. This change adds allowance to the seed's apiserver via the
corresponding network policy.