Use dynamic base id for apiserver-proxy to allow running multiple envoy proxy processes.
Envoy uses shared memory to be able to reattach to the same session during hot restarts.
This requires statically setup identifiers so that the same shared memory segments are
used. It is possible that this clashes if more than one envoy-proxy runs on the same
node. According to https://github.com/envoyproxy/envoy/issues/88, running more than one
envoy-proxy per node is not recommended.
Now, we try to use a dynamic base id for apiserver-proxy as we do not use hot restarts
anyway.