Improve the ObservabilityComponentsHealthy condition for the transition phase
As noted in the PR review by Ismail, the changes in the care/checker.go in the
commit "Adjust the care controller's health check" are not graceful:

> Is this graceful? I don't think it is. On our landscapes Shoots are reconciled
> once in 24h. It means that when the new gardenlet is updated, the health check
> will fail that the grafana deployment is missing until the Shoot is reconciled
> in its maintenance window (when the new grafana Deployment will be deployed).
> You can check for the new grafana Deployment only when the Shoot is reconciled
> with the required gardenlet version.

This commit improves the health check logic of the monitoring components.
Depending on the gardenlet version of the shoot cluster, if it is <1.64, i.e.
before this PR, the former `grafana-{operators,users}` deployments are checked.
For shoots that have already reconciled with gardenlet >=1.64 (i.e. after this
PR), the new `grafana` deployment is checked.

Note: to test this behavior in the local setup, make sure that the gardenlet
configuration property `controllers.shoots.reconcileInMaintenanceOnly` is set to
`true`, to prevent shoot cluster reconciliation when the gardenlet process is
restarted. This way we can create a shoot with an older version of the
gardenlet, restart the gardenlet in the new version and verify that the shoot
condition of the not yet reconciled shoot with the health check of the new
gardenlet is still healthy.

This commit can be reverted in a future release to clean up the code that is
needed for the transition phase after the Gardenlet is updated to 1.64 and
before all the shoots are reconciled in the seed.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>
Co-authored-by: Ismail Alidzhikov <i.alidjikov@gmail.com>