Use a single dash for the ingress prefix of Grafana: "gu-" vs. "gu--" (#7454)
* Use a single dash for the ingress prefix of Grafana: "gu-" vs. "gu--"

Before https://github.com/gardener/gardener/pull/7408#discussion_r1092008535
the ingress domain for the user's Grafana had a "gu-" prefix, e.g.

  gu-local--local.ingress.local.seed.local.gardener.cloud

As a side effect of the change in
https://github.com/gardener/gardener/pull/7408/files#diff-5f6f1c2ea9e5643e60056fa6812709bc31edf03af4752897a694d3fed1f208adR439
the ingress domain changed to:

  gu--local--local.ingress.local.seed.local.gardener.cloud

i.e. had a prefix "gu--" instead of "gu-".

The intention of that change was that the technical ID of new shoots should
consistently use 2 dashes after the prefix "shoot".

This commit adjusts the ComputeIngressHost function to chop off the "shoot--?"
prefix when computing the ingress host name to restore the previous behavior of
the "gu-" prefix. It supports the historic case with the technical ID prefix
"shoot-" of old shoots and the now common case with the technical ID prefix
"shoot--".

Currently the "gu-" prefix (and the other prefixes for Prometheus and the alert
manager) are hard coded in the Gardener Dashboard
https://github.com/gardener/dashboard/blob/1bff671380286345a6d2228a73306f3d524e4a3f/frontend/src/store/modules/shoots/index.js#L185-L190
so changing the prefix would require adjusting the Dashboard links and the links
would be broken until the shoot cluster is reconciled. A better approach could
be to add all the monitoring links (besides Grafana also the link to Prometheus
and to the Alert manager) to the `<shoot>.monitoring` secret so that the
Dashboard does not need to hard code the said prefixes.

With this PR, the original behavior (prefix with a single dash) is restored and
the Dashboard does not need to be adopted.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>

* Address PR review: Merge the ComputeIngressHost test tables

The function shootpkg.ComputeTechnicalID returns the stored TechnicalID
if it exists in the shoot status. This allows us to inject a historic
TechnicalID with a single dash to the existing table test.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>

---------

Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>