Add response rewrite to dns-search-path-optimization. (#7478)
* Add response rewrite to dns-search-path-optimization.

As some clients/DNS resolver validate the hostname in a DNS response,
the hostname in the answer should be rewritten.

As the namespace is not known in all cases, we do not rewrite the
following requests:

* app.namespace.svc.namespace.svc.cluster.local. In this case the 3rd entry
  in the seach-path will match anyway.
* custom.example.com.default.svc.cluster.local. This request will not result
  in answer. However the next case in the search path
  (custom.example.svc.cluster.local) can be rewritten.

* Don't rewrite 'namespace.cluster.local'.

We can't rewrite the answer in the case the search path contains the
source namespace. Hence, we accept that the first query will fail and
rewrite the next search path entry.

* Fix unit tests.

* Allow a leading dot in suffix, but remove the dot.

* Fix linter issue.

* A suffix must contain at least one non-leading dot.