Adapt `NetworkPolicy`s for components in shoot namespaces (#7515)
* Move `kube-apiserver` constants into dedicated package

to prevent cyclic import dependencies

* [`vpn-seed-server`] Replace custom egress rule to `kube-apiserver`

* [`vpn-seed-server`] Replace custom ingress rule from `kube-apiserver`

* [`vpn-seed-server`] Replace custom ingress rule from `istio-ingressgateway`

* Replace `to-shoot-apiserver` usage with `to-kube-apiserver-tcp-443`

Since extensions must be adapted as well, the network policy can just be deprecated but not deleted yet.

* [`gardener-resource-manager`] Replace custom ingress rule from `kube-apiserver`

* [`vpa-admission-controller`] Replace custom ingress rule from `kube-apiserver`

* [`kube-apiserver`] Replace custom egress rule to `etcd`

* [`kube-apiserver`] Replace custom ingress rule from world

* [`grafana`] Replace custom egress rule to `prometheus`

* [`grafana`] Replace custom egress rule to `loki`

* [`prometheus`] Drop egress rule to `istio-ingressgateway`

The Prometheis in shoot namespaces do not scrape the `istio-ingressgateway` - hence, no need to allow egress traffic to it.

* Drop wrong/useless ingress rule from `nginx-ingress` in `kube-system`

The `nginx-ingress` controller runs in the `garden` namespace and not in `kube-system`.

* Keep deprecated `networking.gardener.cloud/to-shoot-apiserver` label for `zero-downtime-validator` job

The upgrade e2e tests always run the `./hack/ci-kind-e2e-upgrade.sh`
script from the HEAD of the tested PR. Hence, the executed e2e test code
in the `./test` directory does not necessarily match with the
environment of the previous release.

* Address PR review feedback