Add validation to the `ManagedSeed` and `ShootManagedSeed` admission plugins
This commit adds validation for the ManagedSeed resource that validates the Shoot `enableStaticTokenKubeconfig` and Seed `secretRef` fields. It is now forbidden to set the Seed `secretRef` when the Shoot static token kubeconfig is disabled. It is also forbidden to disable the Shoot static token kubeconfig when the Seed `secretRef` is set.