Adapt `NetworkPolicy`s for components in `Istio`-related namespaces (#7570)
* Active NetworkPolicy controller for Istio

Active NetworkPolicy controller for Istio namespaces `istio-ingress` and
`istio-system`.

* Replace `Istio-System` netpols

* Allow Istio-Ingress to all shoot apiservers

* Add from-world to Istio-Ingress(es)

Earlier, ingress connections to Istio-Ingress were not restricted. This is not true anymore since the NetworkPolicy controller automatically creates ingress policies for the Istio-Ingress service. Netpols work in a way that once they select a pod, explicit allow listing must happen.

* Run migration for all relevant CP services

Add required annotations for shoot control-plane services once during Gardenlet startup to ensure all services immediately get the expected annotations, independently from the shoot reconciliation. This way, we can delete replaced `NetworkPolicies` in the same release.

* Add 'allow-from-aggregate-prometheus' policy

This policy is needed for Istio-Ingress as long as the
`NetworkPolicy` controller is not activated for the `garden` namespace.
The same policy is already in place for Istiod.

* Use specific annotations for shoot KAS only

* Add test after rebase

* Add label selector for exposureclass handlers

* Address minor review comments

* Determine shoot namespace during runtime

* Handle `deny-all` rule in Gardenlet

* Use ExposureClass name label

Label `handler.exposureclass.gardener.cloud/name` is a common label for Istio-Ingress namespaces that are created for exposure classes.

* Remove VPN shoot rule handling

* Address review comment