Remove no longer necessary Grafana deletion functions
This removes the logic that deletes the deprecated users and operator
Grafana instances. These functions were required for the Grafana
unification change to clean up the instances deployed prior to Gardener
v1.64. With the transition to 1.64, this should have been executed in the
reconciliation of the shoots and the old Grafana instances are expected to
have been deleted.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>