Fix a minor regression in the "remote local setup" scenario (#7618)
* Fix the order of $SUDO and environment variable assignment

If the user is root, the SUDO variable will be empty and the following snippet:

```
SUDO=
$SUDO foo=bar ls
```

will fail with

```
-bash: foo=bar: command not found
```

man bash, "Simple Commands":

> A simple command is a sequence of optional variable assignments followed by
> blank-separated words and redirections, and terminated by a control operator.
> The first word specifies the command to be executed, and is passed as argument
> zero. The remaining words are passed as arguments to the invoked command.

Although `$SUDO` is resolved to blank, the variable assignment has to precede it.

If we switch the order, both cases work as expected and the environment variable
is set on the program that comes after $SUDO:

```
$ SUDO=
$ foo=bar $SUDO env | grep foo
foo=bar
```

```
SUDO="sudo -E"
$ foo=bar $SUDO env | grep foo

Password:
foo=bar
```

Note that when the local setup is executed locally, the user is typically not
root. When it is executed in the "remote local setup" pod, the user is root.

The "remote local setup" case is fixed with this change.

* Address review comments: Move the environment variable to the others