Disable the static token kubeconfig in the shoot template used for TM integration tests (#7609)
* Use `UniversalDeserializer` instead of `UniversalDecoder`

The decoder returned from `UniversalDeserializer` does not perform client-side defaulting, while the decoder returned from `UniversalDecoder` performs client-side defaulting. As the Shoot template under `test/framework/resources/templates/default-shoot.yaml` is partial template, applying defaulting to such kind of partial Shoot template can lead to unexpected behaviour. For example, the K8s version is not set in this template file and it gets set a later point. Applying defaulting while deserializing the object from the template can lead unexpected default values to be set - as the K8s version is missing, any defaulting that is relying on the K8s version to be set will now work as expected.

* Disable the static token kubeconfig in the shoot template used for TM integration tests

In this way we let the gardener-apiserver default the `.spec.kubernetes.enableStaticTokenKubeconfig` field. For Shoots with K8s < 1.26, it will be defaulted to true, for Shoots with K8s >= 1.26 - to false. The plan is to lock this field to false for Shoots with K8s >= 1.27.