Clean up Grafana Unification helper functions (#7615)
* Revert "Improve the ObservabilityComponentsHealthy condition for the transition phase"

This reverts commit edc81d219cf4592351ebb8b557e291760f672dc4.
This logic was needed for the transition phase during which the gardenlet
was updated, but the shoots weren't yet reconciled.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>

* Remove no longer necessary Grafana deletion functions

This removes the logic that deletes the deprecated users and operator
Grafana instances. These functions were required for the Grafana
unification change to clean up the instances deployed prior to Gardener
v1.64. With the transition to 1.64, this should have been executed in the
reconciliation of the shoots and the old Grafana instances are expected to
have been deleted.

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>
Co-authored-by: Jeremy Rickards <jeremy.rickards@sap.com>

---------

Co-authored-by: Istvan Zoltan Ballok <istvan.zoltan.ballok@sap.com>