Adjust Seed Authorizer
Adjust Seed Authorizer to accept list and watch request for various object kinds. When Gardenlet opens a list/watch for single objects, the authorizer can extract the namespace and name from the request and consult the graph to find a relation between the seed and the object in question.
In addition, this commit removes the special case for the "cluster-identity" ConfigMap from the authorizer code. This scenario is now considered when the graph is built.

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>