`Worker` state controller uses PATCH instead of PUT
Earlier, the `Worker` state controller in the extensions library was using PUT/update requests in order to persist the worker state into the `.status.state` field. This led to the fact that null-ed fields were not removed:

```yaml
state:
    machineDeployments:
      shoot--local--local-local:
        machineSets:
        - apiVersion: machine.sapcloud.io/v1alpha1
          kind: MachineSet
          metadata:
            annotations:
              deployment.kubernetes.io/desired-replicas: "1"
              deployment.kubernetes.io/max-replicas: "2"
              deployment.kubernetes.io/revision: "1"
            creationTimestamp: null
            ...
```

However, the `ShootState` extensions sync controller in `gardenlet` was using PATCH requests to copy this data into the `ShootState`. With PATCH requests, the null-ed fields were removed:

```yaml
  spec:
    extensions:
    - kind: Worker
      name: local
      state:
        machineDeployments:
          shoot--local--local-local:
            machineSets:
            - apiVersion: machine.sapcloud.io/v1alpha1
              kind: MachineSet
              metadata:
                annotations:
                  deployment.kubernetes.io/desired-replicas: "1"
                  deployment.kubernetes.io/max-replicas: "2"
                  deployment.kubernetes.io/revision: "1"
                labels:
```

This always led to undesired PATCH requests since the gardenlet's `ShootState` extensions controller considered this a diff.

Now, both controllers use PATCH, hence null-ed fields are removed and the content is in sync in both the `Worker`'s `.status.state` as well as the `ShootState`'s `.spec.extensions[].state` fields.
Unfortunately, we have to reset the state once to ensure currently persisted null-ed fields are removed (the new PATCH call won't do it automatically).