Create a unified logging service definition in the logging stack  (#7623)
* Logging service duplicates the current loki service with the purpose of
providing the same target in fluent-bit outputs.
Once the service is made available on all clusters (clusters are reconciled on a daily basis) the
fluent-bits output will be switched.
After the switch the fluent-bit will deliver logs to the same target logging solving the problem
loki to vali migration. Since the target in fluent-bit is the same, fluent-bit will deliver logs to both
migrated and non migrated shoots.

* Add cleaning up of the newly introduced logging service. It shall be deleted when the logging stack is removed.