Enhance client cache support in Gardenlet (#7632)
* Enable Seed Authorizer for local setup

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>

* Move Aggregator cache to separate file(s)

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>

* Add SingleObject cache

A SingleObject cache is an implementation of the cache.Cache interface which is exclusively responsible for a single object, separated by their object keys (<namespace>/<name>).

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>

* Adjust Seed Authorizer

Adjust Seed Authorizer to accept list and watch request for various object kinds. When Gardenlet opens a list/watch for single objects, the authorizer can extract the namespace and name from the request and consult the graph to find a relation between the seed and the object in question.
In addition, this commit removes the special case for the "cluster-identity" ConfigMap from the authorizer code. This scenario is now considered when the graph is built.

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>

* Enable cache for missing kinds

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>

* Address review comments

* Address review comments II

* Cosmetics and improved logging

* Use `WithOffset` in unit tests

* Use context from `Start` method when starting caches

Earlier, we were using the contexts we received from the `Get()` calls to start the caches. However, those contexts might be very short-lived and are typically canceled too fast, effectively stopping the caches and leading to stale data.

Co-Authored-By: Oliver Götz <47362717+oliver-goetz@users.noreply.github.com>

* Ensure cache is started

Co-Authored-By: Oliver Götz <47362717+oliver-goetz@users.noreply.github.com>

---------

Co-authored-by: Rafael Franzke <rafael.franzke@sap.com>
Co-authored-by: Oliver Götz <47362717+oliver-goetz@users.noreply.github.com>